Retry command execution on concurrency error
Retry executing a command if appending events to the aggregate's stream fails with an `{:error, :wrong_expected_version}` concurrency error. The missing event(s) are fetched from the event store to update the aggregate state and attempt the command again.

Aggregate instances are singletons on a single node, and cluster if using the distributed registry. However, this allows Commanded to still work when run on multiple nodes that are not connected to form a cluster.